Hide plotting behind a simple interface
